Add "Infos" section to sidebar and create info form page
Added a new "Infos" item to the sidebar with a link to the "/infos" route. Created a new composable `useInfoForm.js` to handle form state and validation for the info page. Also, created a new page `index.vue` under `app/pages/infos` to display the form for entering a resume URL.